requirements.txt
pandas==1.2.4
psycopg2-binary==2.9.1
pymongo==3.12.0